get, add, edit and delete implemented

+++ src/main/java/br/com/challenge/swapi/controllers/dtos/PlanetDTO.java
@@ -13,5 +13,5 @@ public class PlanetDTO {
     private String name;
     private String climate;
     private String terrain;
-    private Long appearances;
+    private Integer appearances;
 }
